pyproject.toml
[tool.poetry]
name = "utterances-filter"
version = "0.1.0"
description = ""
authors = ["Jean-Luc Thirion <thirion.jl@gmail.com>"]
readme = "README.md"
packages = [
    { include = "app" },
    { include = "dataload" }
]

[tool.poetry.dependencies]
python = "^3.10"
nltk = "^3.8.1"
pyyaml = "^6.0"
scikit-learn = "^1.2.0"
sentence-transformers = "^2.2.2"
h5py = "^3.8.0"

[tool.poetry.group.ml.dependencies]
matplotlib = "^3.6.3"
jupyter = "^1.0.0"
seaborn = "^0.12.2"
pandas = "^1.5.3"
tensorboard = "^2.11.2"
hydra-core = "^1.3.1"
svgling = "^0.3.1"
bokeh = "^3.0.3"
llvmlite = [
    { version = "^0.39.1", markers = "platform_machine == 'arm64'" }
]
umap-learn = "^0.5.3"

[tool.poetry.group.app.dependencies]
fastapi = "^0.89.0"
uvicorn = { extras = ["standard"], version = "^0.20.0" }
python-multipart = "^0.0.5"
aiofiles = "^22.1.0"
requests = "^2.28.2"
websockets = "^10.4"
pyaudio = "^0.2.13"

[tool.poetry.group.scrape.dependencies]
beautifulsoup4 = "^4.11.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
black = {version = "^23.1.0", extras = ["d"]}
watchdog = "^2.3.0"
mypy = "^1.0.1"


[tool.poetry.group.backoffice.dependencies]
streamlit = "^1.19.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # The following are specific to Black, you probably don't want those.
  | blib2to3
  | tests/data
  | profiling
)/
'''
